Some comments in code; most are pretty obvious, tbh.

+++ loader/stage3/screen.cpp
@@ -1,6 +1,8 @@
 #include "screen.h"
 #include "physmem.h"
 
+// helper functions
+// in and out on a port
 void outb(short port, char value)
 {
     asm volatile ("outb %1, %0" : : "dN" (port), "a" (value));
@@ -19,11 +21,14 @@ namespace Screen
 
     void Initialize()
     {
+        // initialize kout
+        // YES, I know that this call is not the prettiest one
         Screen::kout = (Screen::Console *)PhysMemory::Manager::Place(sizeof(Screen::Console));
         Screen::kout->Initialize();
     }
 }
 
+// setups default values for members
 void Screen::Console::Initialize()
 {
     this->m_iMaxX = 80;
@@ -39,54 +44,71 @@ void Screen::Console::Initialize()
 
 Screen::Console * Screen::Console::Print(char iChar)
 {
+    // newline encountered
     if (iChar == '\n')
     {
+        // go to new line
         this->m_iY++;
         this->m_iX = 0;
         
+        // scroll if necessary
         if (this->m_iY == this->m_iMaxY)
         {
             this->Scroll();
         }
     }
     
+    // tabulator
     else if (iChar == '\t')
     {
+        // find how many spaces it is needed to fill the tabulator
         int i = (this->m_iX + 1) % 8;
         if (i == 0)
             i = 8;
         
+        // and print them
         while (i-- != 0)
         {
             this->Print(' ');
         }
     }
     
+    // carriage return; simplest possible
     else if (iChar == '\r')
     {
         this->m_iX = 0;
     }
     
+    // backspace
     else if (iChar == '\b')
     {
+        // overwrite the previous char
         this->m_iX--;
         this->Print(' ');
+        // and set position at it
         this->m_iX--;
     }
     
+    // actual printing
     else
     {
+        // find where in memory the char should be written
         volatile char * pos = this->m_pScreenMemory;
         pos += 2 * (this->m_iY * this->m_iMaxX + this->m_iX);
+        // write character 
         *(pos) = iChar;
+        // and attribute
         *(pos + 1) = this->m_iAttrib;
         
         this->m_iX++;
+        // end of line encountered
         if (this->m_iX == this->m_iMaxX)
         {
+            // new line
             this->m_iX = 0;
             this->m_iY++;
             
+            // scroll if necessary
             if (this->m_iY == this->m_iMaxY)
             {
                 this->Scroll();
@@ -94,6 +116,7 @@ Screen::Console * Screen::Console::Print(char iChar)
         }
     }
     
+    // update cursor position
     this->MoveCursor();
     
     return this;
@@ -103,9 +126,11 @@ Screen::Console * Screen::Console::Print(char * pString)
 {
     char c;
     
+    // print every character in string
     while (true)
     {
         c = *pString;
+        // and stop at null
         if (c == 0)
             break;
         
@@ -123,6 +148,7 @@ Screen::Console * Screen::Console::Print(long int iInt)
     switch (this->m_eMode)
     {
         case iBin:
+            // binary printing
             for (int i = 63; i >= 0; i--)
             {
                 if (iInt & (1 << i))
@@ -137,6 +163,7 @@ Screen::Console * Screen::Console::Print(long int iInt)
             
             break;
         case iOct:
+            // octal printing
             tmp = iInt / 8;
             mod = iInt % 8;
             
@@ -147,6 +174,7 @@ Screen::Console * Screen::Console::Print(long int iInt)
             
             break;
         case iDec:
+            // decimal printing
             tmp = iInt / 10;
             mod = iInt % 10;
             
@@ -157,6 +185,7 @@ Screen::Console * Screen::Console::Print(long int iInt)
             
             break;
         case iHex:
+            // hex printing
             const char * digits = "0123456789ABCDEF";
             
             for (int i = 60; i > -1; i -= 4)
@@ -170,11 +199,12 @@ Screen::Console * Screen::Console::Print(long int iInt)
 
 Screen::Console * Screen::Console::Print(long int * pInt)
 {
-    return this->Print(reinterpret_cast<long int>(pInt));
+    return this->Print(reinterpret_cast<int>(pInt));
 }
 
 Screen::Console * Screen::Console::Print(int iInt)
 {
+    // see long int version
     int tmp, mod;
     
     switch (this->m_eMode)
@@ -238,16 +268,17 @@ Screen::Console * Screen::Console::Print(double fDouble)
 
 Screen::Console * Screen::Console::Print(double * pDouble)
 {
-    return this->Print(reinterpret_cast<long int>(pDouble));
+    return this->Print(reinterpret_cast<int>(pDouble));
 }
 
 Screen::Console * Screen::Console::Print(void * pPointer)
 {
-    return this->Print(reinterpret_cast<long int>(pPointer));
+    return this->Print(reinterpret_cast<int>(pPointer));
 }
 
 void Screen::Console::MoveCursor()
 {
+    // move cursor; assembly magic
     asm("pusha");
 
     short i = this->m_iY * this->m_iMaxX + this->m_iX;
@@ -264,7 +295,8 @@ void Screen::Console::Scroll()
 {
     volatile char * target = this->m_pScreenMemory;
     volatile char * base = this->m_pScreenMemory + this->m_iMaxX * 2;
-    
+
+    // move everything one line up
     for (int i = 0; i < this->m_iMaxX * (this->m_iMaxY - 1) * 2; i++)
     {
         *target = *base;
@@ -276,6 +308,7 @@ void Screen::Console::Scroll()
     this->m_iX = 0;
     this->m_iY = this->m_iMaxY - 1;
     
+    // print spaces at last line
     for (int i = 0; i < this->m_iMaxX; i++)
     {
         this->Print(' ');
@@ -289,12 +322,14 @@ void Screen::Console::Clear()
 {
     volatile char * base = this->m_pScreenMemory;
     
+    // clear
     for (int i = 0; i < this->m_iMaxX * this->m_iMaxY; i++)
     {
         base[2 * i] = ' ';
         base[2 * i + 1] = Screen::Console::m_iAttrib;
     }
     
+    // gotoxy(0, 0);
     this->m_iX = 0;
     this->m_iY = 0;
     this->MoveCursor();
